+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Add 'include_common_names' parameter to API Otu autocomplete
 - DwC import task: Detect and highlight columns that will be ignored [#4406]
 - Project data export task: Recent download table [#4470]
+- Field synchronize: Object label column [#4477]
 
 ### Fixed
 
@@ -58,6 +59,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 [#4466]: https://github.com/SpeciesFileGroup/taxonworks/issues/4466
 [#4470]: https://github.com/SpeciesFileGroup/taxonworks/issues/4470
 [#4472]: https://github.com/SpeciesFileGroup/taxonworks/issues/4472
+[#4477]: https://github.com/SpeciesFileGroup/taxonworks/issues/4477
 
 ## [0.52.1] - 2025-06-17